services_backup_20250524_234954/search/vector-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.0
numpy==1.24.3
faiss-cpu==1.7.4
pgvector==0.2.4
openai==1.3.7
httpx==0.25.2 